+++ components/calendar/index.ts
@@ -1,0 +1,1 @@
+export { TrainingCalendar, TrainingCalendarSkeleton } from "./TrainingCalendar";